+++ src/__tests__/se.ts
@@ -1,0 +1,16 @@
+import {
+    doneTests,
+    initTests,
+} from "./init";
+import { runCommand } from "../controllers";
+import { consoleTerminal } from "../core/utils";
+
+beforeAll(initTests);
+afterAll(doneTests);
+
+test("Use custom docker image", async () => {
+    await runCommand(consoleTerminal, "se set", {
+        instance: "se1",
+        image: "tonlabs/local-node:0.28.1",
+    });
+});