+++ jest.config.js
@@ -1,0 +1,11 @@
+module.exports = {
+    "roots": [
+        "<rootDir>/src",
+    ],
+    "testMatch": [
+        "**/__tests__/**/*.+(ts|tsx|js)",
+    ],
+    "transform": {
+        "^.+\\.(ts|tsx)$": "ts-jest",
+    },
+};