+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.8.0] - 2021-07-26
+
+### New
+- `value` option of the `contract` commands can accept values with "t" suffix.
+  Such values will be properly converted to nano tokens.
+- `run-signer` option of the the `contract` commands allows to sign messages with alternative keys different from `signer` meanwhile address is still calculated from `signer` parameter (or from default `signer`). It should be useful for `run-local` (with `None` value) so that it is not needed to specify address explicitly any more.
+- `tondev` commands that produce files (e.g. `sol create`) create output folders if required.
+- `tondev signer add` command's `secret` option can accept a path to the keys file.
+- `tondev se set` options `image` and `container` allow use existing docker 
+  images and containers to register SE instances.
+- `tondev se delete` deletes registered SE instance from `tondev` SE registry.
+- `--version` or `-v` or `-V` global option prints current tondev version.
+
+### Fixed
+- "tondev clang set --compiler" didn't change installed compiler version https://github.com/tonlabs/tondev/issues/42
+- `tondev` did not not fill the command args with default values when user ran commands programmatically.  
+- `tondev info` stopped if one of the tools failed.
+
+### Improved
+- In case of errors with giver `tondev` prints new detailed errors describing 
+  that there is a giver problem (not in users contract). 
+- Component version table prints "not installed" text if some component is missing. Footnote about on demand installation is added.
+- Account balance check is added before deploy.
+- All commands that require abi file (e.g. `js wrap`) now accepts any input file name. 
+- `tondev se version` now prints version from the latest to the eldest order.
+- `tondev sol compile` now hides linker output in case of success.
+
 ## [0.7.4] - 2021-06-25
 
 ### New